edit product tests added
