+++ SadnaSrc/StoreCenterTests/StoreCenterTests.csproj
@@ -66,7 +66,7 @@
   <ItemGroup>
     <Compile Include="AddProductsTests.cs" />
     <Compile Include="AddToCartTests.cs" />
-    <Compile Include="EditProduct.cs" />
+    <Compile Include="EditProductTests.cs" />
     <Compile Include="RemoveProductTests.cs" />
     <Compile Include="OpenStoreTests.cs" />
     <Compile Include="StoreDL TEST.cs" />
